% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.timeseries.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{A Plotting function used to plot peaks in a time series}
\usage{
\method{plot}{timeseries}(data, file, sample = NULL, time.start = NULL, time.stop = NULL)
}
\arguments{
\item{data}{output from the timeseries.peaks() function}

\item{file}{the name of the file the user wishes to plot}

\item{sample}{the specific sample the user wishes to plot}

\item{time.start}{option to specify when you want the plot to start}

\item{time.stop}{option to specify when you want the plot to stop}
}
\description{
This function provides a plotting framework for exploring the data output from the timeseries.peaks() function
}
\examples{
setwd(path.package("peak.gas"))
test <- timeseries.peaks()
# File wide examples
Plot.timeseries(test, file = "vn_veg_07292021.txt")
Plot.timeseries(test, file = "vn_veg_07292021.txt", time.start = "2021-08-02 13:30:00")
Plot.timeseries(test, file = "vn_veg_07292021.txt", time.stop = "2021-08-02 12:40:00")
Plot.timeseries(test, file = "vn_veg_07292021.txt", time.start = "2021-08-02 12:45:00", time.stop = "2021-08-02 13:00:00")
# Specific Sample examples
Plot.timeseries(test, file = "vn_veg_07292021.txt", sample = "NS_Veg")
Plot.timeseries(test, file = "vn_veg_07292021.txt", sample = "NS_Veg", time.start = "2021-08-02 13:01:00")
Plot.timeseries(test, file = "vn_veg_07292021.txt", sample = "NS_Veg", time.stop = "2021-08-02 12:54:00")
Plot.timeseries(test, file = "vn_veg_07292021.txt", sample = "NS_Veg", time.start = "2021-08-02 12:56:00", time.stop = "2021-08-02 13:01:00")
}
\keyword{CO2,}
\keyword{Li-Cor,}
\keyword{LiCor,}
\keyword{Peak,}
\keyword{gas,}
\keyword{plot}
