% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.peaks.R
\name{timeseries.peaks}
\alias{timeseries.peaks}
\title{A function to combine and name timeseries text files}
\usage{
timeseries.peaks()
}
\description{
This function will loop through a folder containing text file outputs to combine and assign sample names, outputing a single timeseries
}
\examples{
setwd(path.package("peak.gas"))
output <- timeseries.peaks()
}
\keyword{CO2,}
\keyword{Li-Cor,}
\keyword{LiCor,}
\keyword{Peak,}
\keyword{gas,}
\keyword{plot}
