% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.extracted.R
\name{plot.extracted}
\alias{plot.extracted}
\title{A Plotting function used with output from CO2 Peak finding function}
\usage{
\method{plot}{extracted}(data, file = NULL, sample = NULL, std.curve = FALSE, method = "linear")
}
\arguments{
\item{data}{output from the extract.peaks() function}

\item{file}{the name of the file the user wishes to look at}

\item{sample}{the specific sample the user wishes to look at}

\item{std.curve}{a logical argument that states whether the user wishes to look at the standard curve or not}

\item{method}{Specify whether you want to plot the "linear" relationship or the "log" transformed relationship}
}
\description{
This function provides a plotting framework for exploring the data output from the extract.peaks() function
}
\examples{
ex.output <- extract.peaks(directory = path.package("peak.gas"))
plot(ex.output, std.curve = TRUE)
plot(ex.output, file = "vn_clear_071621.txt", std.curve = TRUE)
plot(ex.output, file = "vn_clear_07292021.txt", std.curve = TRUE)
plot(ex.output, file = "vn_veg_07292021.txt", std.curve = TRUE)
plot(ex.output, file = "vn_darkveg_071621.txt")
plot(ex.output, sample = "NN_DARKVEG")
plot(ex.output, file = "vn_darkveg_071621.txt", sample = "NC_DARKVEG")
}
\keyword{CO2,}
\keyword{Li-Cor,}
\keyword{LiCor,}
\keyword{Peak,}
\keyword{gas,}
\keyword{plot}
